software/fdcan-gui/requirements.txt
PySide6==6.5.0
pyserial
